django_todos/todos_app/serializers.py
from rest_framework import serializers

from todos_app.models import Todo


class TodoModelSerializer(serializers.ModelSerializer):
    class Meta:
        model = Todo
        fields = ('id', 'title', 'completed')
