requirements.txt
Django==2.0.4
requests==2.18.4
django-model-utils==3.1.1
djangorestframework==3.8.2
django-extensions==2.2.5
 django-model-utils==3.2.0
ipython==6.3.0
